Implemented report_id and SQLite provider
