+++ kai/config.toml
@@ -2,6 +2,24 @@ log_level = "info"
 demo_mode = false
 trace_enabled = true
 
+# **Postgresql incident store**
+# [incident_store]
+# provider = "postgresql"
+
+# [incident_store.args]
+# host = "127.0.0.1"
+# database = "kai"
+# user = "kai"
+# password = "dog8code"
+
+# **In-memory sqlite incident store**
+# ```
+# [incident_store]
+# provider = "sqlite"
+#
+# [incident_store.args]
+# connection_string = "sqlite:///:memory:"
+
 [incident_store]
 provider = "postgresql"
